rename sym mem buffer file
