+++ torchao/prototype/moe_training/ep/syncless/__init__.py
@@ -1,4 +1,4 @@
-from torchao.prototype.moe_training.ep.syncless.buffer_manager import (
+from torchao.prototype.moe_training.ep.syncless.sym_mem_buffer_manager import (
     SymmetricMemoryBufferManager,  # noqa: F401
     get_buffer_manager,  # noqa: F401
 )
